AxiomAI/tiny_gpt/requirements.txt
torch>=2.0.0
tqdm
numpy
Pillow
torchvision
